Fix tests for KubernetesDeployments
